+++ projectApp/tests/unit_tests/__mocks__/MockFile.ts
@@ -1,0 +1,8 @@
+// Mock File class
+export default class {
+    constructor(
+      public fileBits: BlobPart[],
+      public name: string,
+      public options?: FilePropertyBag,
+    ) {}
+  };
