+++ projectApp/tests/unit_tests/__mocks__/mockFileReader.ts
@@ -1,6 +1,7 @@
 import { sendError } from "../../../src/logger-frontend";
+import { readFile } from "fs/promises";
 
-export class MockFileReader {
+export default class MockFileReader {
   result: string | null = null;
   onload: ((event: ProgressEvent<FileReader>) => void) | null = null;
   onerror: ((event: ProgressEvent<FileReader>) => void) | null = null;
@@ -24,5 +25,21 @@ export class MockFileReader {
     }
   }
 }
-
-export default { MockFileReader };
+ /**
+* Returns a file that has the contents of the file at the filePath
+* @param filePath path string to the file
+* @returns `File` {fileBits: contents at filePath, name: filePath, options: in text/csv}
+*/
+export async function pathStrToFile(filePath: string): Promise<File> {
+ const reader = readFile(filePath, "utf-8");
+ let data: string;
+ try {
+   data = await reader;
+ } catch {
+   data = "";
+   //for some reason, if await reader throws the error, vitest still receives ENOENT error
+   //vitest cant handle errors outside of expect
+   //so instead data will be returned as empty (aka couldnt read)
+ }
+ return new File([data], filePath, { type: "text/csv" });
+}
